Extract expiry logic into dedicated module and fix conventions
- Extract FALLBACK_EXPIRY_OPTIONS, SHORTHAND_MAP, getServerConfig,
  and parseExpiry from cli.js into new expiry.js module
- cli.js now 499 lines (under 500-line limit)
- Add PHPDoc block to ConfigController::__invoke
- Fix misleading JSDoc on fetchConfig (no longer a public endpoint)
